Factor out ref utilities from RemoteButler
Pull out some standalone functions from RemoteButler to a new file to cut down clutter.
